+++ src/content-routing/index.js
@@ -10,9 +10,9 @@ const Query = require('../query')
 const utils = require('../utils')
 
 /**
- * @typedef {import('cids')} CID
+ * @typedef {import('multiformats/cid').CID} CID
  * @typedef {import('peer-id')} PeerId
- * @typedef {import('multiaddr')} Multiaddr
+ * @typedef {import('multiaddr').Multiaddr} Multiaddr
  */
 
 /**
@@ -98,6 +98,7 @@ module.exports = (dht) => {
 
       provs
         .forEach(id => {
+          /** @type {{ id: PeerId, addresses: { multiaddr: Multiaddr }[] }} */
           const peerData = dht.peerStore.get(id)
 
           if (peerData) {