+++ src/peer-list/peer-distance-list.js
@@ -2,9 +2,9 @@
 
 const utils = require('../utils')
 const pMap = require('p-map')
-const uint8ArrayEquals = require('uint8arrays/equals')
-const uint8ArrayCompare = require('uint8arrays/compare')
-const uint8ArrayDistance = require('uint8arrays/xor')
+const { equals: uint8ArrayEquals } = require('uint8arrays/equals')
+const { compare: uint8ArrayCompare } = require('uint8arrays/compare')
+const { xor: uint8ArrayXor } = require('uint8arrays/xor')
 
 /**
  * @typedef {import('peer-id')} PeerId
@@ -56,7 +56,7 @@ class PeerDistanceList {
     const dhtKey = await utils.convertPeerId(peerId)
     const el = {
       peerId,
-      distance: uint8ArrayDistance(this.originDhtKey, dhtKey)
+      distance: uint8ArrayXor(this.originDhtKey, dhtKey)
     }
 
     this.peerDistances.push(el)
@@ -83,7 +83,7 @@ class PeerDistanceList {
     const furthestDistance = this.peerDistances[this.peerDistances.length - 1].distance
 
     for (const dhtKey of dhtKeys) {
-      const keyDistance = uint8ArrayDistance(this.originDhtKey, dhtKey)
+      const keyDistance = uint8ArrayXor(this.originDhtKey, dhtKey)
 
       if (uint8ArrayCompare(keyDistance, furthestDistance) < 0) {
         return true