+++ test/query/index.spec.js
@@ -2,10 +2,7 @@
 /* eslint max-nested-callbacks: ["error", 6] */
 'use strict'
 
-const chai = require('chai')
-chai.use(require('dirty-chai'))
-chai.use(require('chai-checkmark'))
-const expect = chai.expect
+const { expect } = require('aegir/utils/chai')
 const sinon = require('sinon')
 const delay = require('delay')
 const PeerStore = require('libp2p/src/peer-store')
@@ -17,7 +14,7 @@ const c = require('../../src/constants')
 const createPeerId = require('../utils/create-peer-id')
 const { sortClosestPeers } = require('../../src/utils')
 const { convertBuffer } = require('../../src/utils')
-const uint8ArrayFromString = require('uint8arrays/from-string')
+const { fromString: uint8ArrayFromString } = require('uint8arrays/from-string')
 const NUM_IDS = 101
 
 describe('Query', () => {