+++ test/rpc/handlers/find-node.spec.js
@@ -1,12 +1,10 @@
 /* eslint-env mocha */
 'use strict'
 
-const chai = require('chai')
-chai.use(require('dirty-chai'))
-const expect = chai.expect
+const { expect } = require('aegir/utils/chai')
 const Message = require('../../../src/message')
 const handler = require('../../../src/rpc/handlers/find-node')
-const uint8ArrayFromString = require('uint8arrays/from-string')
+const { fromString: uint8ArrayFromString } = require('uint8arrays/from-string')
 
 const T = Message.TYPES.FIND_NODE