+++ test/rpc/handlers/get-value.spec.js
@@ -1,13 +1,11 @@
 /* eslint-env mocha */
 'use strict'
 
-const chai = require('chai')
-chai.use(require('dirty-chai'))
-const expect = chai.expect
+const { expect } = require('aegir/utils/chai')
 const Message = require('../../../src/message')
 const handler = require('../../../src/rpc/handlers/get-value')
 const utils = require('../../../src/utils')
-const uint8ArrayFromString = require('uint8arrays/from-string')
+const { fromString: uint8ArrayFromString } = require('uint8arrays/from-string')
 
 const T = Message.TYPES.GET_VALUE